CPU: Fix NMI timing bug
When an NMI is requested to the CPU, it is first asserted at the end of
the current instruction, and then the interrupt sequence starts when the
next instruction finishes.

This fixes a bug where some games hung at the title screen since the NMI
triggered one instruction too early.