Netplay: Implement basic lockstep synchronization
Button presses are "commited" (i. e. notified to the emulator) at the
start of each frame. Besides, the commited button state is sent to the
other host over the network.

None of the hosts' emulators is allowed to progress until they have sent
their button states and received those of their respective peers.